Run extra new tests during the aegir3-drush-master-dev test.

+++ tests/functional_tests/aegir3-drush-master-dev/run-tests.sh
@@ -12,5 +12,12 @@ then
   exit 1
 fi
 
+vagrant ssh -c "sudo -u root -s /bin/bash -c 'su aegir -l -c \"drush -y @hostmaster --debug provision-tests-new-run\"'"
+if [ "$?" -ne "0" ]
+then
+  echo "'provision-tests-run' failed. Leaving vm in place for forensic analysis."
+  exit 1
+fi
+
 vagrant destroy --force
 exit 0
